requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
google-generativeai
python-dotenv
pydantic 